update ui of several pages
