+++ assets/css/cards.scss
@@ -839,7 +839,6 @@ $wildcard-mobile-color: rgb(251, 182, 124);
   .name {
     margin-right: 5px;
     margin-top: 5px;
-    text-align: center;
     line-height: 16px;
   }
 
@@ -917,8 +916,8 @@ $wildcard-mobile-color: rgb(251, 182, 124);
   font-size: 10px;
   margin: 0;
   text-indent: -9999em;
-  width: 2em;
-  height: 2em;
+  width: 8em;
+  height: 8em;
   border-radius: 50%;
   background: #7a7a7a;
   background: -moz-linear-gradient(left, #7a7a7a 10%, rgba(122,122,122, 0) 42%);
