+++ lib/copi_web/components/layouts/app.html.heex
@@ -1,5 +1,5 @@
 <div class="min-h-full">
-  <div class="bg-gray-800 pb-32">
+  <div class="bg-gray-800">
     <nav class="bg-gray-800">
       <div class="mx-auto max-w-7xl sm:px-6 lg:px-8">
         <div class="border-b border-gray-700">
@@ -10,7 +10,7 @@
                   <!-- Current: "bg-gray-900 text-white", Default: "text-gray-300 hover:bg-gray-700 hover:text-white" -->
                   <a href="/" class="bg-gray-900 text-white rounded-md px-3 py-2 text-sm font-medium" aria-current="page">Home</a>
                   <a href="/games/new" class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Start A Game</a>
-                  <a href="#" class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Join A Game</a>
+                  <%!-- <a href="#" class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Join A Game</a> --%>
                   <a href="/cards" class="text-gray-300 hover:bg-gray-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Game Cards</a>
                 </div>
               </div>
@@ -46,16 +46,11 @@
         </div>
       </div>
     </nav>
-    <header>
-      <%!-- <div class="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
-        <h1 class="text-3xl font-bold tracking-tight text-white">Welcome to OWASP Copi!</h1>
-      </div> --%>
-    </header>
   </div>
 
-  <main class="-mt-32">
+  <main class="">
     <div class="mx-auto max-w-7xl px-4 pb-12 sm:px-6 lg:px-8">
-      <div class="rounded-lg bg-white px-5 py-6 shadow sm:px-6">
+      <div class="h-auto bg-white rounded-lg shadow sm:px-6">
         <.flash_group flash={@flash} />
         <%= @inner_content %>
       </div>
