+++ lib/copi_web/components/layouts/root.html.heex
@@ -28,7 +28,7 @@
       </script>
     <% end %>
   </head>
-  <body class="h-full">
+  <body>
     <%= @inner_content %>
   </body>
 </html>
