+++ lib/copi_web/controllers/card_html/index.html.heex
@@ -1,5 +1,3 @@
-<h1>Cornucopia Card Deck</h1>
-
-<%= for card <- @cards do %>
-  <CopiWeb.CardHTML.show card={card} />
-<% end %>
+  <%= for card <- @cards do %>
+      <CopiWeb.CardHTML.show card={card} />
+  <% end %>
