+++ lib/copi_web/live/game_live/index.ex
@@ -1,6 +1,7 @@
 defmodule CopiWeb.GameLive.Index do
   use CopiWeb, :live_view
   use Phoenix.Component
+  
   import CopiWeb.LiveHelpers
 
   alias Copi.Cornucopia
