+++ lib/copi_web/live/player_live/form_component.html.heex
@@ -1,5 +1,5 @@
 <div>
-  <h2><%= @title %></h2>
+  <.header1><%= @title %></.header1>
 
   <.form :let={f} for={@changeset} id="player-form" phx-target={@myself} phx-change="validate" phx-submit="save">
